+++ source/extensions/omni.flux.info_icon.widget/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0]
+### Changed
+- Allow specifying the icon size
+
 ## [1.1.4]
 ### Fixed
 - Fixed tests flakiness